Deal with long for getSecondsDisplayed(), since when viewing large time periods >38 years, the seconds overlaps.
git-svn-id: file:///C/dev/jaret/timebars/svnrepo/branches/1.48changes@22 49765bf9-81a4-5f4f-8ba1-75f15dd23e95

+++ src/main/java/de/jaret/util/ui/timebars/TimeBarViewerInterface.java
@@ -1083,7 +1083,7 @@ public interface TimeBarViewerInterface {
      * 
      * @return the number of seconds currently displayed by the diagram geometry
      */
-    int getSecondsDisplayed();
+    long getSecondsDisplayed();
 
     /**
      * Retrieve the used strategy for determing overlap information.
